stop targeting bash shell specifically:
I think this script was written before MacOS (OSX) changed the default
shell from bash to zsh, thus the first versions of this script were
based on the default shell at the time. Only just now did it occur to me
that this is not necessary for modern MacOS.

However, I have older machines that would still benefit from
this script. I will likely aim for general `sh` compatibility and try to
avoid any 'zsh-isms' in an effort to support down to MacOS 10.13 (High Sierra).